requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
croniter==5.0.1
cryptography==43.0.3
debtcollector==3.0.0
decorator==5.1.1
distlib==0.3.8
dnspython==2.7.0
dogpile.cache==1.3.3
eval_type_backport==0.2.0
eventlet==0.38.0
fasteners==0.19
filelock==3.13.3
Flask==3.1.0
Flask-Session==0.8.0
frozenlist==1.5.0
futurist==3.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.9.0
keystonemiddleware==10.8.0
kombu==5.4.2
logutils==0.3.5
lxml==5.1.0
Mako==1.3.6
MarkupSafe==3.0.2
mistral==19.0.0
mistral-lib==3.0.0
mistralai==1.2.3
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
mypy-extensions==1.0.0
netaddr==1.3.0
networkx==3.4.2
openai==0.27.2
os-service-types==1.7.0
oslo.cache==3.9.0
oslo.concurrency==6.2.0
oslo.config==9.7.0
oslo.context==5.7.0
oslo.db==17.0.0
oslo.i18n==6.5.0
oslo.log==6.2.0
oslo.messaging==15.0.0
oslo.metrics==0.10.0
oslo.middleware==6.3.0
oslo.policy==4.5.0
oslo.serialization==5.6.0
oslo.service==3.6.0
oslo.utils==7.4.0
osprofiler==4.2.0
packaging==24.2
paramiko==3.5.0
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.0
pecan==1.5.1
platformdirs==4.2.0
ply==3.11
prettytable==3.12.0
prometheus_client==0.21.0
propcache==0.2.0
psutil==6.1.0
pycadf==4.0.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-keystoneclient==5.5.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
repoze.lru==0.7
requests==2.32.3
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.21.0
setuptools==75.6.0
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
statsd==4.0.1
stevedore==5.4.0
tenacity==9.0.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tooz==6.3.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
virtualenv==20.25.1
voluptuous==0.15.2
wcwidth==0.2.13
WebOb==1.8.9
Werkzeug==3.1.3
wrapt==1.17.0
WSME==0.12.1
yappi==1.6.10
yaql==3.0.0
yarl==1.18.0
zipp==3.21.0
